requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# Web Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# HTML Generation
jinja2>=3.1.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Data Validation
pydantic>=2.0.0

# Utilities
aiohttp>=3.9.0

# Rich console for beautiful terminal output
rich>=13.7.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# TLS Fingerprinting Bypass (opcional)
# Para habilitar bypass TLS fingerprinting, instalar con:
# pip install curl-cffi>=0.6.0
# curl-cffi>=0.6.0
